Letters after periods in U.R.L.s no longer are no longer capitalized by the AutoCorrect system.
Also, /say is now (properly) handled by myScribe with a default (not yet configurable) epithet for the console ("&dGod") and AutoCorrections on the messages themselves.

Also, I found out that color codes still act like punctuation when it comes to spacing, which causes unnecessary spacing where color codes are in the final message.